# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Koperasi Sekar Kartini API",
        "description": "Documentation API Koperasi Sekar Kartini",
        "contact": {
            "name": "API Support",
            "email": "support@koperasisekarkartini.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "password"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "example": "081200000001"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcdefghijklmnopqrstuvwxyz"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Account not verified or not active"
                    }
                }
            }
        },
        "/api/register-group-member": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new group member",
                "operationId": "ea24ea28b6209179017f10bf91202e0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "member_number",
                                    "identity_number",
                                    "birth_date",
                                    "phone_number",
                                    "address",
                                    "occupation",
                                    "identity_card_photo",
                                    "self_photo",
                                    "password",
                                    "work_area_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Budi Santoso"
                                    },
                                    "member_number": {
                                        "description": "Nomor anggota unik",
                                        "type": "string",
                                        "example": "MBR-001"
                                    },
                                    "identity_number": {
                                        "description": "NIK unik",
                                        "type": "string",
                                        "example": "3201010101900001"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "phone_number": {
                                        "description": "Nomor telepon unik",
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Raya No. 123"
                                    },
                                    "occupation": {
                                        "type": "string",
                                        "example": "Petani"
                                    },
                                    "identity_card_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "self_photo": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "work_area_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registration successful. Please wait for verification."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Logged out successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Group": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "number": {
                        "type": "string",
                        "example": "GRP001"
                    },
                    "description": {
                        "type": "string",
                        "example": "Kelompok tani sejahtera",
                        "nullable": true
                    },
                    "shared_liability_fund_amount": {
                        "type": "integer",
                        "example": 50000
                    },
                    "group_fund_amount": {
                        "type": "integer",
                        "example": 30000
                    },
                    "social_fund_amount": {
                        "type": "integer",
                        "example": 20000
                    },
                    "total_shared_liability_fund": {
                        "type": "integer",
                        "example": 500000
                    },
                    "total_group_fund": {
                        "type": "integer",
                        "example": 300000
                    },
                    "total_social_fund": {
                        "type": "integer",
                        "example": 200000
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "work_area_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "chairman_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "facilitator_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "treasurer_id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Rahmat Admin"
                    },
                    "phone_number": {
                        "type": "string",
                        "example": "081200000001"
                    },
                    "member_number": {
                        "type": "string",
                        "example": "ADM-001"
                    },
                    "identity_number": {
                        "type": "string",
                        "example": "3201111111110001"
                    },
                    "address": {
                        "type": "string",
                        "example": "Jl. Anggur No. 10, Sumbersari, Jember"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "is_verified": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "work_area": {
                        "$ref": "#/components/schemas/WorkArea"
                    },
                    "group": {
                        "$ref": "#/components/schemas/Group"
                    }
                },
                "type": "object"
            },
            "WorkArea": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name_work_area": {
                        "type": "string",
                        "example": "Wilayah Kerja A"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use your API token as Bearer token for authorization",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}